fix(transaction): update CRUD arg type
